Add database integration
Database integration code has been cleaned up so that Models may be added more easily.
The database is imported into the app and set to 'db' for use by the rest of the application.
Note: the database is imported asynchronously

A basic User ORM has been created which allows the addition, modification, and deletion of Users from the database.

Gulpfile:
 - rename 'servercreate' to 'serverinit'
 - serverinit calls 'serverdelete' first
 - rename 'serversetup' to 'serverconfig'

Dependencies:
 - bcrypt
 - valid-email