Adds user signup and login functionality
Adds user signup functionality

- Adds db models into server requests
We included the db models object into every request received by
the server. This way, we can access the bookshelf models from
every request handler we desire.

- Fixes a bug in the User.setPassword method
There was a `this` closure created before an inner function was
called in this methods. One of the lines inside this inner function
used `this` as opposed to `_this` <- closure

- Adds user signup functionality
We added the user save functionality in the POST /auth/signup requests
handler. Sends back a 201 and a sessionID cookie when the user is created,
or a `409 Conflict` when the user already exists.

Add user login functionality

- Includes the `cors` express library
Since our mobile app is not being hosted in the same domain as our server,
we need to handle Cross Origin Resource Sharing headers. The `cors`
express library handles this really well.

- Fixes problem with user signup
There was a small bug with user signups. There was no response if the user
had an invalid password. This is now fixed.

- Adds user login functionality
Users are now able to login.

- TODO:
Session handling

Fix: Adds new line at end of login-router